Moved remaining collector handlers to app/handler/
