+++ cmd/collector/app/handler/tchannel_handler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package app
+package handler
 
 import (
 	"github.com/uber/tchannel-go/thrift"
