+++ cmd/collector/app/handler/thrift_span_handler.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package app
+package handler
 
 import (
 	"go.uber.org/zap"
